Adding removal of metrics from reporters for carbon and console
This adds a new method on the console and carbon reporters
that can be used to remove a metric from their records.

This will be used in the following cases:
 - Metrics on dynamic objects
 - Control of metrics so that you can turn on and off metrics programatically.

The end goal is to have this as a method on Reporter, but I thought I would
get feedback on carbon and console before implementing the more tricky prometheus
metrics.